refac: consolidated all platform specifics to single pairs of files
